Refactor: extract `MembersControls` from `ColonyMembers`
