+++ src/modules/dashboard/components/ColonyMembers/ColonyMembers.css.d.ts
@@ -3,4 +3,3 @@ export const mainContentGrid: string;
 export const mainContent: string;
 export const rightAside: string;
 export const loadingWrapper: string;
-export const controls: string;
