+++ src/modules/dashboard/components/ColonyMembers/MembersControls/MembersControls.css
@@ -1,0 +1,3 @@
+.controls li + li {
+  margin: 18px 0 0;
+}
