+++ src/modules/dashboard/components/ColonyMembers/MembersControls/MembersControls.css.d.ts
@@ -1,0 +1,1 @@
+export const controls: string;
