+++ src/modules/dashboard/components/ColonyMembers/MembersControls/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MembersControls';
